requirements.txt
flask==2.3.3
tiktoken==0.5.2
Werkzeug==2.3.7
pillow==10.0.0 